* combination of scrolling / zooming now works without intervening fixateVisibleImage   - solved by using floats instead of integers when calculating and by correct moving of the pivot points
** now there seems to be an issue with creating the non-filtered bitmap

+++ app/src/main/java/jetsetilly/mandelbrot/Gestures/GestureHandler.java
@@ -1,10 +1,10 @@
 package jetsetilly.mandelbrot.Gestures;
 
 public interface GestureHandler {
-    void checkActionBar(int x, int y, boolean allow_show);
-    void scroll(int x, int y);
+    void checkActionBar(float x, float y, boolean allow_show);
+    void scroll(float x, float y);
     void finishManualGesture();
-    void autoZoom(int offset_x, int offset_y, boolean zoom_out);
+    void autoZoom(float offset_x, float offset_y, boolean zoom_out);
     void manualZoom(float amount);
     void endManualZoom();
 }
